backend/requirements.txt
boto3==1.34.145
botocore==1.34.145
facenet_pytorch==2.5.2
fastapi==0.115.12
numpy==2.2.5
openai==1.75.0
Pillow==11.2.1
pydantic==2.11.3
PyJWT==2.10.1
redis==5.2.1
Requests==2.32.3
SQLAlchemy==2.0.31
torch==2.6.0
torchvision==0.21.0
uvicorn==0.34.2
python-dotenv
psycopg2-binary
python-multipart